lesson 1 code update

+++ Lesson1.HW.2023.09.09/project3/project3.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-	int a;
+	long long a;
 	cin >> a;
 	int b = a / 10;
 	int c = b * (b + 1);
